Fix #135
